refactored GameController to core module
